requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pyautogui==0.9.54
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.4.2
pillow==10.0.1
python-multipart==0.0.6
